Refactor mode-core and discovery-module and readwrite-module dependencie
